pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lintx"
dynamic = ["version"]
description = 'lint agent'
readme = "README.md"
requires-python = ">=3.10,<=3.12"
license = "MIT"
keywords = ["ai", "agent", "lint"]
authors = [
  { name = "Jia Jia", email = "angersax@sina.com" },
]
dependencies = [
  "camel-ai[model_platforms,rag,storage]>=0.2.59",
  "dotenv>=0.9.9",
  "requests_oauthlib>=2.0.0",
]

[project.urls]
Documentation = "https://github.com/ai-agentx/lintx/blob/main/README.md"
Issues = "https://github.com/ai-agentx/lintx/issues"
Source = "https://github.com/ai-agentx/lintx"

[tool.hatch.build.targets.wheel]
packages = ["src/lintx"]
include = ["camel"]

[tool.hatch.build.targets.sdist]
packages = ["src/lintx"]
include = ["camel"]

[tool.hatch.version]
path = "src/lintx/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/lintx tests}"

[tool.coverage.run]
source_pkgs = ["lintx", "tests"]
branch = true
parallel = true
omit = [
  "src/lintx/__about__.py",
]

[tool.coverage.paths]
lintx = ["src/lintx", "*/lintx/src/lintx"]
tests = ["tests", "*/lintx/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
